requirements.txt
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
distlib==0.3.6
fastapi==0.88.0
filelock==3.8.0
h11==0.14.0
idna==3.4
natsort==8.2.0
numpy==1.23.5
pickles==0.1.1
platformdirs==2.5.2
pydantic==1.10.2
python-dotenv==0.21.0
requests==2.28.1
sniffio==1.3.0
starlette==0.22.0
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.20.0
virtualenv==20.16.5
